Se cambio int por double en el metodo Atacar de la interfaz Atacante y se agrego el metodo getVidaActual a la interfaz Atacable

+++ src/interfaces/unidadesYEstructuras/Atacante.java
@@ -1,5 +1,5 @@
 package interfaces.unidadesYEstructuras;
 
 public interface Atacante {
-	public void atacar(Atacable atacable, int distanciaEntreAtacanteYAtacable);
+	public void atacar(Atacable atacable, double distanciaEntreAtacanteYAtacable);
 }
